requirements.txt
tensorflow==2.6.2
opencv-python==4.5.1.48
bokeh==2.2.3
#webbrowser
plotly==4.14.3
shap==0.38.1
pathlib
#base64
#time
seaborn==0.11.0
matplotlib==3.3.2
Pillow==8.0.1
numpy==1.19.5
pandas==1.1.3
streamlit_player==0.1.5
streamlit==1.3.1